requirements.txt
# =============================================================================
# JOBPER v4.0 — CRM de Contratos para Colombia
# =============================================================================
# pip install -r requirements.txt
# Para CPU-only torch: pip install torch --index-url https://download.pytorch.org/whl/cpu
# =============================================================================

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# Base de datos
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Auth
PyJWT>=2.8.0

# Validation
pydantic[email]>=2.5.0

# Cache + Queue
redis>=5.0.0
celery>=5.3.0

# Search
elasticsearch>=8.11.0

# HTTP Client
requests>=2.31.0

# Email (Resend — via requests, no SDK needed)
# Push notifications
pywebpush>=2.0.0

# Scheduler
schedule>=1.2.0

# Variables de entorno
python-dotenv>=1.0.0

# =============================================================================
# NLP / EMBEDDINGS
# =============================================================================
sentence-transformers>=2.2.0
numpy>=1.24.0
torch>=2.0.0

# =============================================================================
# SCRAPERS
# =============================================================================
beautifulsoup4>=4.12.0
lxml>=4.9.0

# =============================================================================
# UTILIDADES
# =============================================================================
python-dateutil>=2.8.0
pytz>=2023.3
bleach>=6.1.0
